Updated product version to 0.9.2 and improved deployment docs
Upgraded all versions used in builds from 0.9.1 to 0.9.2. Added support
for versions in docker-compose.yml as well as in applicable docker-*.sh
commands.  Updated README.md docs and deployment documentation to match
current steps including use of versions.

+++ services/goldfin-build/docker-down.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
 # Take down composed containers.
+export VERSION=$1
+if [ "$VERSION" = "" ]; then
+  echo "Usage $0 <version>"
+  exit 1
+fi
 set -x
 docker-compose -p goldfin down
